Rename `kits` export to `legacy.js`
